Added Collisions and spawnpoint to map3
Part 10 of Limeoats cavestory tutorial series.
Player can no longer walk through walls. This was done by adding collision objects to map3. A rectangle class was created and added to the sprite class as boundingBox. When the boundingBox of the player overlaps with a collision object on the map, the collision side is computed, then the player is pushed back. In-game it simply appears as if the player hits the wall.
A spawnPoint has been added to map3.