Fixes to tileset code in level class
Tileset struct now has a LastGid int to help with loading the correct tileset in level.cpp. Previously the first tileset would be chosen for each gid, as the only criteria was that the tilesets first gid is <= the current tiles gid. Now the current tiles gid must be in betweent the first and last gid of the tileset, ensuring proper selection.
Tilesets after the first one will have FirstGid values greater than 1. This will confuse tileset positioning during LoadMap, so gid is reduced by FirstGid for each tile.
Tileset positioning was incorrect for tiles at the end of the tileset. -1 subtraction was occuring at the wrong time for the tsxx component and not at all for the tsyy component.
Redundancy reduced in animatedsprite.cpp by calling stopAnimation in update.