Minor changes
Framerate calculation fixed. Contained unnecessary multiplication.
Change to the doorway in the street map.
Shooting animation fixed. Previously, the first frame of the animation did not have the player firing. This meant that if the user clicked fast enough, the animation would end before it got to the firing frame, which appeared laggy.