Movement between levels added
New door class was added. These are objects in the level similar to collisions.
Each game loop, the level checks to see if the players boundingBox is colliding with a door (Level::checkDoorCollisions). If there is a collision, this door object is passed to Player::handleDoorCollisions, along with the current Level and Graphics. handleDoorCollisions takes the destination from the Door object to change the map.
Level::getPlayerSpawn has been edited for this new functionality. It now has an optional argument 'from'. If this argument is included, Level will check its spawnPointMap for a key of value 'from' and use this spawn point if it exists. Otherwise it will just use the spawn point with key 'default'.
A new map 'apartment' was added for this feature. There are also some changes to map 'street'.