Sight mechanic added
Line of sight between the player and the enemies is now updated each frame. This is done with Bresenham's line algorithm. A line is drawn between the player and the enemy. Each point is checked against the collisionRects. If the line is not obstructed, the enemy has its inSight field set to true.
inSight field is now used in Alien class. If player is within sight and range, the alien will give chase. If line of sight is lost, the alien will continue to the last seen location of the player.